==> Pubmed_Bert/1.yn_prepare.ipynb <==
import json
import pandas as pd
import numpy as np
import os
import string
import eval
from mkdir import mk_dir
from bert_format import bset, train_yn, test_yn

# 训练集

label_path = "/home/data/t070224/Pubmed/Pubmed_Label/3.Orgnize/output/文献合并"
col_dic = {"是否是致病菌":"pat","免疫状态":"immu","预后":"oc"}

for col_name, col_ab in col_dic.items():
    train_qa = []
    # 读取标注信息
    df = pd.read_excel("%s/human_label_pat.xlsx" % (label_path), header=0)
    
    # 设置问题
    if col_name == "是否是致病菌":
        que = []
        for idx in df.index:
            que.append('Can %s cause pneumonia?' % df.loc[idx, "物种名称"])
    if col_name == "免疫状态":
        df = df.loc[df["是否是致病菌"] == "yes", :]
        df.loc[df[col_name] == "抑制", col_name] = "yes"
        que = 'Is patient immunocompromised?'
    if col_name == "预后":
        df = df.loc[df["是否是致病菌"] == "yes", :]
        df.loc[df[col_name] == "死亡", col_name] = "yes"
        que = 'Is patient dead?'
    
    df.loc[df[col_name] != "yes", col_name] = "no"    
    df["question"] = que
    
    # 创建QA样本  
    for idx in df.index:
        train_qa.append(train_yn(
            '%s:%s' % (df.loc[idx, "ID"], df.loc[idx, "物种名称"]),
            df.loc[idx, "question"],
            '%s' % df.loc[idx, col_name],
            '%s' % df.loc[idx, "Full"]))
    bset(train_qa, "input/input_%s/"%(col_ab), "train_set.json")

#  测试集

with open("/home/data/t070224/Pubmed/Pubmed_Search/search_results/search_results_hanimals.txt","r") as f:
    content = f.read()
    spec_papers_dic = json.loads(content)

# 预测样本格式转换
papers_l = {}
for spec_type in spec_papers_dic:

    for spec,spec_papers in spec_papers_dic[spec_type].items():
        
        for paper_id, paper in spec_papers.items():
            
            papers_l['%s:%s' % (paper_id, spec)] = paper


test_qa = []
for paper_id,paper in papers_l.items():     
    if len(paper["texts"]["Ab"]) > 0:
        test_qa.append(
            test_yn(
                paper_id, 
                'Can %s cause Pneumonia?' % paper_id.split(":")[-1],
                paper["texts"]["Ab"]+" "+" ".join(sum(paper["Meshhead"],[]))))
bset(test_qa,"input/input_pat/","test_qa.txt")
