==> Pubmed_Bert/2.collect_yn.ipynb <==
import numpy as np
import pandas as pd
import json
import os

# Merge

file_list = [i for i in os.listdir("output/output_pat/test_set") if "test_set" in i]
prediction = {}
for folder in file_list:
    with open("output/output_pat/test_set/%s/predictions_.json" % folder, "r") as f:
        prediction.update(json.load(f))

prediction_yes = {}
for k,v in prediction.items():
    if v[0] == "yes":
        prediction_yes[k] = v

with open("output/output_pat/total_predictions.json", "w") as f:
    json.dump(prediction_yes, f)
