==> Pubmed_Bert/4.collect_qa.ipynb <==
# import package

import matplotlib.pyplot as plt
plt.rcParams['font.sans-serif']=['SimHei']
import os
import numpy as np
import pandas as pd
import json
from copy import deepcopy
import re
import shutil
import random
import string
from random import choice
random.seed(10)
# -*- coding:utf-8 -*
import eval
from mkdir import mk_dir
from bert_format import bset,test_qa
import transfer

# transfer num

col_ab = "num"
col_name = "样本量"

prediction_0 = pd.DataFrame(np.zeros((0,1)), columns = [col_ab])
for re in range(0,6):
    
    prediction = json.load(open("output/output_%s/output_%s/predictions_.json" % (col_ab, re),'r'))
    prediction = pd.DataFrame(prediction, index=[col_ab]).T
    prediction_0 = pd.concat((prediction_0, prediction),axis=0)

prediction_num = transfer.trans_num(prediction_0, col_ab)

# transfer age

col_ab = "age"
col_name = "年龄"

prediction_0 = pd.DataFrame(np.zeros((0,1)), columns = [col_ab])
for re in range(0,6):
    
    prediction = json.load(open("output/output_%s/output_%s/predictions_.json" % (col_ab, re),'r'))
    prediction = pd.DataFrame(prediction, index=[col_ab]).T
    prediction_0 = pd.concat((prediction_0, prediction),axis=0)

prediction_age = transfer.trans_age(prediction_0, col_ab)

# transfer gender

col_ab = "gender"
col_name = "患者性别"

prediction_0 = pd.DataFrame(np.zeros((0,1)), columns = [col_ab])
for re in range(0,6):
    
    prediction = json.load(open("output/output_%s/output_%s/predictions_.json" % (col_ab, re),'r'))
    prediction = pd.DataFrame(prediction, index=[col_ab]).T
    prediction_0 = pd.concat((prediction_0, prediction),axis=0)

prediction_gender = transfer.trans_gender(prediction_0, col_ab)

# transfer immu

col_ab = "immu"
col_name = "免疫状态"

prediction_0 = pd.DataFrame(np.zeros((0,1)), columns = [col_ab])
for re in range(0,6):
    
    prediction = json.load(open("output/output_%s/output_%s/nbest_predictions_.json" % (col_ab, re),'r'))
    pred_l = {}
    for k, item in prediction.items():
        pred_l[k] = [item[0]["text"],item[0]["probability"]]
    prediction = pd.DataFrame(pred_l, index=[col_ab, "probability"]).T
    prediction_0 = pd.concat((prediction_0, prediction),axis=0)
prediction_immu = transfer.trans_immu(prediction_0[[col_ab]], col_ab)

# transfer evi

col_ab = "evi"
col_name = "证据等级"

prediction_0 = pd.DataFrame(np.zeros((0,1)), columns = [col_ab])
for re in range(0,6):
    
    prediction = json.load(open("output/output_%s/output_%s/predictions_.json" % (col_ab, re),'r'))
    prediction = pd.DataFrame(prediction, index=[col_ab]).T
    prediction_0 = pd.concat((prediction_0, prediction),axis=0)

# prediction_evi = transfer.transfer_evi(prediction_0, col_ab)

prediction_0.to_csv("output/样本类型预测文本.csv")

# 合并数据

prediction_addtion = pd.concat([prediction_num,prediction_age,prediction_gender,prediction_immu, prediction_evi], axis=1)

prediction_addtion.to_csv("output/prediction_addition.csv")
